# ollantaweb/docs/swagger.yaml
basePath: /
definitions:
...
  api.tokenListResponse:
    properties:
      tokens:
        items:
          $ref: '#/definitions/api.tokenView'
        type: array
    type: object
  api.tokenView:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      last_used_at:
        type: string
      name:
        type: string
      project_id:
        type: integer
      token_type:
        type: string
    type: object
...
  api.userListResponse:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/api.userView'
        type: array
    type: object
  api.userView:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      login:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
  api.webhookTestResponse:
    properties:
      status:
        type: string
    type: object
...
  postgres.Webhook:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: Ollanta REST API documentation.
  title: Ollanta API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Returns paginated list of users
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userListResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new local user
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            login:
              type: string
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.userView'
      summary: Create user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Deactivate a user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Deactivate user
      tags:
      - users
    get:
      description: Returns a single user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userView'
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user name, email, or avatar
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: body
        required: true
        schema:
          properties:
            avatar_url:
              type: string
            email:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userView'
      summary: Update user
      tags:
      - users
  /api/v1/users/{id}/reactivate:
    post:
      description: Reactivate a deactivated user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Reactivate user
      tags:
      - users
  /api/v1/users/{id}/tokens:
    get:
      description: List API tokens for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.tokenListResponse'
      summary: List user tokens
      tags:
      - users
  /api/v1/users/{id}/tokens/{tid}:
    delete:
      description: Delete a specific token for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token ID
        in: path
        name: tid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete user token
      tags:
      - users
...
  /api/v1/users/me/password:
    put:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password data
        in: body
        name: body
        required: true
        schema:
          properties:
            new_password:
              type: string
            old_password:
              type: string
          type: object
      responses:
        "204":
          description: No Content
      summary: Change password
      tags:
      - users
...
  /api/v1/webhooks:
    get:
      description: Returns webhooks, optionally filtered by project
      parameters:
      - description: Project key
        in: query
        name: project_key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/postgres.Webhook'
            type: array
      summary: List webhooks
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Create a new webhook
      parameters:
      - description: Webhook data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/postgres.Webhook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/postgres.Webhook'
      summary: Create webhook
      tags:
      - webhooks
  /api/v1/webhooks/{id}:
    delete:
      description: Delete a webhook
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete webhook
      tags:
      - webhooks
    put:
      consumes:
      - application/json
      description: Update a webhook
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      - description: Webhook data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/postgres.Webhook'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/postgres.Webhook'
      summary: Update webhook
      tags:
      - webhooks
...
  /api/v1/webhooks/{id}/test:
    post:
      description: Fire a test event to a webhook
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.webhookTestResponse'
      summary: Test webhook
      tags:
      - webhooks
...
swagger: "2.0"